+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Added recovery method reminder email, replacing daily method verify emails
 
 ## [4.1.1] - 2019-04-18
 ### Fixed